Kill old listening thread when a new one arrives.
This ensures (almost, there's a race condition) that when a user
refreshes their page, they are promptly sent the root items (locations
if they're a GM, their own character if they're a player (once I
implement that)). Even if it does fail, it's just a 30s wait for
the next request anyway.