Add player character items
Spawn player characters in as items, updating the client records as
players join too. Improve the item consistency checks, adding more
checks and only checking those attributes that have updated. Add a
failsafe timeout so that if an update to the world takes more than
1s to complete, the server gives up and leaves the world with its
previous state. (This should be a 500 error but it's currently just
204 No Content. I think this may be an issue with the library.)